projects/mditations_rag/requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
pypdf2==3.0.1
chromadb==0.4.22
sentence-transformers==2.2.2
google-generativeai==0.3.2
streamlit==1.29.0
python-dotenv==1.0.0
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3 